# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "EndpointIn": {
        "properties": {
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          }
        },
        "required": [
          "endpoint"
        ],
        "title": "EndpointIn",
        "type": "object"
      },
...
      "EventOut": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delivered": {
            "title": "Delivered",
            "type": "boolean"
          },
          "dismissed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dismissed At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "title",
          "body",
          "delivered",
          "dismissed_at",
          "created_at"
        ],
        "title": "EventOut",
        "type": "object"
      },
...
      "PushKeyOut": {
        "properties": {
          "devices": {
            "title": "Devices",
            "type": "integer"
          },
          "public_key": {
            "title": "Public Key",
            "type": "string"
          }
        },
        "required": [
          "public_key",
          "devices"
        ],
        "title": "PushKeyOut",
        "type": "object"
      },
      "PushSubscriptionIn": {
        "properties": {
          "auth": {
            "title": "Auth",
            "type": "string"
          },
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "p256dh": {
            "title": "P256Dh",
            "type": "string"
          }
        },
        "required": [
          "endpoint",
          "p256dh",
          "auth"
        ],
        "title": "PushSubscriptionIn",
        "type": "object"
      },
...
      "RuleIn": {
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "bot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bot Id"
          },
          "cooldown_minutes": {
            "default": 0,
            "title": "Cooldown Minutes",
            "type": "integer"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "once": {
            "default": false,
            "title": "Once",
            "type": "boolean"
          },
          "params": {
            "additionalProperties": true,
            "default": {},
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "kind"
        ],
        "title": "RuleIn",
        "type": "object"
      },
      "RuleOut": {
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "bot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bot Id"
          },
          "bot_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bot Name"
          },
          "cooldown_minutes": {
            "title": "Cooldown Minutes",
            "type": "integer"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "kind_label": {
            "title": "Kind Label",
            "type": "string"
          },
          "last_fired_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Fired At"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "once": {
            "title": "Once",
            "type": "boolean"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "id",
          "kind",
          "kind_label",
          "params",
          "account_id",
          "bot_id",
          "bot_name",
          "note",
          "enabled",
          "once",
          "cooldown_minutes",
          "last_fired_at"
        ],
        "title": "RuleOut",
        "type": "object"
      },
...
      "TelegramIn": {
        "properties": {
          "chat_id": {
            "title": "Chat Id",
            "type": "string"
          }
        },
        "required": [
          "chat_id"
        ],
        "title": "TelegramIn",
        "type": "object"
      },
      "TelegramOut": {
        "properties": {
          "bot_configured": {
            "title": "Bot Configured",
            "type": "boolean"
          },
          "chat_id": {
            "title": "Chat Id",
            "type": "string"
          },
          "uses_default_chat": {
            "title": "Uses Default Chat",
            "type": "boolean"
          }
        },
        "required": [
          "chat_id",
          "uses_default_chat",
          "bot_configured"
        ],
        "title": "TelegramOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionAuth": {
        "in": "cookie",
        "name": "sessionid",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "",
    "title": "Trade desk API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/alerts/defaults": {
      "post": {
        "operationId": "alerts_api_defaults",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Defaults",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/events": {
      "get": {
        "description": "`active`: only alerts not yet dismissed, the ones the app keeps on screen.",
        "operationId": "alerts_api_list_events",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "default": false,
              "title": "Active",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "List Events",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/events/dismiss-all": {
      "post": {
        "operationId": "alerts_api_dismiss_all",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Dismiss All",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/events/{event_id}/dismiss": {
      "post": {
        "operationId": "alerts_api_dismiss",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Dismiss",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/push": {
      "get": {
        "operationId": "alerts_api_push_status",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushKeyOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Push Status",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/push/subscribe": {
      "post": {
        "operationId": "alerts_api_push_subscribe",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushKeyOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Push Subscribe",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/push/test": {
      "post": {
        "operationId": "alerts_api_push_test",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Push Test",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/push/unsubscribe": {
      "post": {
        "operationId": "alerts_api_push_unsubscribe",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushKeyOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Push Unsubscribe",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/rules": {
      "get": {
        "operationId": "alerts_api_list_rules",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "List Rules",
        "tags": [
          "alerts"
        ]
      },
      "post": {
        "operationId": "alerts_api_create_rule",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Create Rule",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/rules/{rule_id}": {
      "delete": {
        "operationId": "alerts_api_delete_rule",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Delete Rule",
        "tags": [
          "alerts"
        ]
      },
      "put": {
        "operationId": "alerts_api_update_rule",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Update Rule",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/telegram": {
      "get": {
        "operationId": "alerts_api_get_telegram",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Get Telegram",
        "tags": [
          "alerts"
        ]
      },
      "put": {
        "operationId": "alerts_api_set_telegram",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Set Telegram",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/telegram/test": {
      "post": {
        "operationId": "alerts_api_test_telegram",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Test Telegram",
        "tags": [
          "alerts"
        ]
      }
    },
...
  },
  "servers": []
}